added Attractor function for indents at specific position

+++ Software/Knobber_STM32/Core/Inc/Motorcontrol.h
@@ -14,7 +14,7 @@
 
 void initMotor();
 void updateMotor();
-void setMotorPower(uint8_t power);
+void setMotorPower(uint16_t power);
 uint16_t getMotorAngle();
 void setMotorAngle(uint16_t angle);
 void setMotorDirection(bool direction);
